% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Batch}
\alias{Batch}
\title{Similarity Batch matrix}
\format{An object of class \code{matrix} with 500 rows and 500 columns.}
\usage{
Batch
}
\description{
This similarity matrix represents batch effects
1 reflects samples from the same batch

See extra_scripts/make_similarity_matrix.R for the script used to
generate the similarity matrix
}
\keyword{datasets}
