% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gmt}
\alias{Gmt}
\title{Similarity DNA methylation matrix}
\format{

}
\usage{
Gmt
}
\description{
This similarity matrix is based on the simulated DNA
methylation dataset cpg_simulated

See extra_scripts/make_similarity_matrix.R for the script used to
generate the similarity matrix
}
\keyword{datasets}
