% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pheno_similarity.R
\name{match_pheno_similarity}
\alias{match_pheno_similarity}
\title{Match phenotype and similarity matrix participant order by ID variables}
\usage{
match_pheno_similarity(phenotype, similarity, id)
}
\arguments{
\item{phenotype}{A phenotype data.frame.}

\item{similarity}{A similarity matrix.}

\item{id}{A string for ID variable in phenotype data \code{x},
which corresponds to \code{rownames} in similarity matrix \code{y}}
}
\value{
A list of matched phenotype and similarity matrix data.
}
\description{
Match phenotype and similarity matrix participant order by ID variables
}
\examples{
# Load datasets
data("phenotype") # Phenotype data
data("Gmt")       # Methylation similarity matrix

# Match phenotype and methylation similarity matrix by ID variable
matched.list <- match_pheno_similarity(phenotype, Gmt, "ID")

# Save the data as separate objects again
phenotype_matched <- matched.list[[1]]
Gmt_matched <- matched.list[[2]]
}
