% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsR2.R
\name{omicsR2}
\alias{omicsR2}
\title{Estimate variance explained by a similarity matrix above
variance explained by covariates}
\usage{
omicsR2(outcome, fixed_covar, random_full, random_baseline,
  validation_proportion = 0.2, repetitions = 100, seed = 0)
}
\arguments{
\item{outcome}{A string specifying the outcome variable name.}

\item{fixed_covar}{A string specyfying the fixed effects.}

\item{random_full}{A list of similarity matrices to test.}

\item{random_baseline}{A list of similarity matrices to adjust for.}

\item{validation_proportion}{The size of the validation set in percentage.}

\item{repetitions}{How often should cross-validation be repeated.}

\item{seed}{Seed to ensure reproducibility of cross-validation.}
}
\value{
Vector of variance explained estimates from cross-validation.
}
\description{
Estimate variance explained by a similarity matrix above
variance explained by covariates
}
\examples{
# Load datasets
data("phenotype") # Phenotype data
data("Gmt")       # Methylation similarity matrix
data("Batch")     # Batch similarity matrix

# Regress outcome on methylation similarity matrix, batch similarity matrix,
# two fixed effect covariates with GREML and Monte-Carlo Cross-Validation.
# Returns variance explained minus the variance explained by fixed effect
# covariates and similarity matrices specified in random_baseline.
Gmt_variance_explained <- omicsR2(outcome = outcome, fixed_covar = "covariate1 + covariate2",
 random_full = list(Methylation=Gmt, Batch=Batch),
  random_baseline = list(Batch=Batch), validation_proportion = 0.2,
  repetitions = 100, seed = 20190405)

# Examine the variance explained distribution from cross-validation
mean(Gmt_variance_explained)
quantile(Gmt_variance_explained)
hist(Gmt_variance_explained)
}
