% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsR2_imp.R
\name{omicsR2_imp}
\alias{omicsR2_imp}
\title{Estimate variance explained by a similarity matrix above
variance explained by covariates for multiply imputed data. This function
runs \code{\link{omicsR2}} for each imputed datasets. Results
are then pooled using Rubin's rule.}
\usage{
omicsR2_imp(
  outcome,
  fixed_covar,
  random_full,
  random_baseline,
  imp_list,
  validation_proportion = 0.2,
  repetitions = 100,
  ncores_imp = 1,
  ncores = 1,
  seed = 0
)
}
\arguments{
\item{outcome}{A string specifying the outcome variable name.}

\item{fixed_covar}{A string specifying the fixed effects.}

\item{random_full}{A list of similarity matrices to test.}

\item{random_baseline}{A list of similarity matrices to adjust for.}

\item{imp_list}{A list of imputed datasets.}

\item{validation_proportion}{The size of the validation set in percentage.}

\item{repetitions}{How often should cross-validation be repeated.}

\item{ncores_imp}{number of cores used to run imputed datasets in parallel.}

\item{ncores}{Number of cores to run cross-validations in parallel.}

\item{seed}{Seed to ensure reproducibility of cross-validation.}
}
\value{
List with three data.frames:
1. Per imputed dataset: data.frame with variance explained estimates
per cross-validation for baseline (r2_covariates) and full model (r2_full),
plus the difference (r2_diff).
2. Per imputed dataset, the mean and sd of the r2 estimates across
cross-validations.
3. The pooled estimates across imputations using Rubin's rule.
}
\description{
Estimate variance explained by a similarity matrix above
variance explained by covariates for multiply imputed data. This function
runs \code{\link{omicsR2}} for each imputed datasets. Results
are then pooled using Rubin's rule.
}
\examples{
# Load datasets
data("phenotype") # Phenotype data
data("Gmt")       # Methylation similarity matrix
data("Batch")     # Batch similarity matrix

# Randomly set covariates to missing
set.seed(20190405)
phenotype_matched_na <- phenotype
phenotype_matched_na$covariate1 <- ifelse(phenotype_matched_na$covariate1 <= quantile(phenotype_matched_na$covariate1, .3), NA, phenotype_matched_na$covariate1)
phenotype_matched_na$covariate2 <- ifelse(phenotype_matched_na$covariate2 <= quantile(phenotype_matched_na$covariate2, .3), NA, phenotype_matched_na$covariate2)

# Impute data five times
miceImps <- mice(phenotype_matched_na)

# Create a list of imputed datasets
impList <- list()
for (i in 1:miceImps$m) impList[[i]] <- complete(miceImps, action = i)

# Regress outcome on methylation similarity matrix, batch similarity matrix,
# two fixed effect covariates with GREML and Monte-Carlo Cross-Validation.
# Returns variance explained by DNA methylation minus the variance explained by
# fixed effect covariates and similarity matrices specified in random_baseline.
# Training:validation ratio is 80:20 and randomly sampled 100 times.
# Results are pooled across the list of imputed datasets.
Gmt_variance_explained <- omicsR2_imp(outcome = "outcome",
                                  fixed_covar = "covariate1 + covariate2",
                                  random_full = list(Methylation=Gmt_matched, Batch=Batch),
                                  random_baseline = list(Batch=Batch),
                                  imp_list = impList,
                                  validation_proportion = 0.2, repetitions = 100, seed = 20190405)

# Examine the variance explained distribution across cross-validations and
# imputed datasets
all_estimates.data <- do.call(rbind, Gmt_variance_explained_imp[[1]])
hist(all_estimates.data$r2_diff, breaks = 10)
Gmt_variance_explained_imp[[3]]
}
