% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{A data.frame with simulated confounders and outcome}
\format{A data.frame of 3 variables for 500 participants
\describe{
  \item{ID}{participant ID}
  \item{covariate1}{Simulated covariate 1}
  \item{covariate1}{Simulated covariate 2}
  \item{outcome}{Simulated outcome}
}}
\usage{
phenotype
}
\description{
This simulated dataset contains two confounders and ane outcome. These
variables were specified to be associated with 5 CpG sites of the
cpg_simulated dataset.

See extra_scripts/simulate_covariates_outcomes.R for the script used to
generate the dataset
}
\keyword{datasets}
