% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gmt}
\alias{Gmt}
\title{Similarity DNA methylation matrix}
\format{An object of class \code{matrix} with 500 rows and 500 columns.}
\usage{
Gmt
}
\description{
This similarity matrix is based on the simulated DNA
methylation dataset cpg_simulated

See extra_scripts/make_similarity_matrix.R for the script used to
generate the similarity matrix
}
\keyword{datasets}
